Move defaultHost into seperate file for mocking.
