+++ packages/structure/src/ide.ts
@@ -1,5 +1,3 @@
-import * as fs from 'fs-extra'
-import glob from 'glob'
 import { basename } from 'path'
 import * as tsm from 'ts-morph'
 import { TextDocuments } from 'vscode-languageserver'
@@ -11,29 +9,17 @@ import {
   Range,
   Hover,
 } from 'vscode-languageserver-types'
-import { Paths, getPaths } from '@redwoodjs/internal'
 import { ArrayLike, ArrayLike_normalize } from './x/Array'
 import { lazy, memo } from './x/decorators'
 import { basenameNoExt } from './x/path'
 import { createTSMSourceFile_cached } from './x/ts-morph'
 import { URL_file } from './x/URL'
 import { ExtendedDiagnostic } from './x/vscode-languageserver-types'
+import { Host, DefaultHost } from './hosts'
+
+export { Host, DefaultHost }
 
 export type NodeID = string
-
-/**
- * The host interface allows us to decouple the "model/*"
- * classes from access to the file system.
- * This is critical for editor support (ex: showing diagnostics on unsaved files)
- */
-export interface Host {
-  existsSync(path: string): boolean
-  readFileSync(path: string): string
-  readdirSync(path: string): string[]
-  globSync(pattern: string): string[]
-  writeFileSync(path: string, contents: string): void
-  paths: Paths
-}
 
 export type IDEInfo =
   | Definition
@@ -246,28 +232,6 @@ export abstract class FileNode extends BaseNode {
   }
   @lazy() get basename() {
     return basename(this.filePath)
-  }
-}
-
-export class DefaultHost implements Host {
-  existsSync(path: string) {
-    return fs.existsSync(path)
-  }
-  readFileSync(path: string) {
-    return fs.readFileSync(path, { encoding: 'utf8' }).toString()
-  }
-  readdirSync(path: string) {
-    return fs.readdirSync(path)
-  }
-  globSync(pattern: string) {
-    return glob.sync(pattern)
-  }
-  writeFileSync(path: string, contents: string) {
-    return fs.writeFileSync(path, contents)
-  }
-  @lazy()
-  get paths() {
-    return getPaths()
   }
 }
